+++ web/src/components/admin/users/buttons/DeactivaterButton.tsx
@@ -1,0 +1,73 @@
+import {
+  type User,
+  UserStatus,
+  UserRole,
+  USER_ROLE_LABELS,
+  INVALID_ROLE_HOVER_TEXT,
+} from "@/lib/types";
+import { type PageSelectorProps } from "@/components/PageSelector";
+import { HidableSection } from "@/app/admin/assistants/HidableSection";
+import { PopupSpec } from "@/components/admin/connectors/Popup";
+import userMutationFetcher from "@/lib/admin/users/userMutationFetcher";
+import useSWRMutation from "swr/mutation";
+import {
+  Table,
+  TableHead,
+  TableRow,
+  TableBody,
+  TableCell,
+} from "@/components/ui/table";
+
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import { Button } from "@/components/ui/button";
+import { GenericConfirmModal } from "@/components/modals/GenericConfirmModal";
+import { useState } from "react";
+import { usePaidEnterpriseFeaturesEnabled } from "@/components/settings/usePaidEnterpriseFeaturesEnabled";
+import { DeleteEntityModal } from "@/components/modals/DeleteEntityModal";
+import { TableHeader } from "@/components/ui/table";
+
+export const DeactivaterButton = ({
+  user,
+  deactivate,
+  setPopup,
+  mutate,
+}: {
+  user: User;
+  deactivate: boolean;
+  setPopup: (spec: PopupSpec) => void;
+  mutate: () => void;
+}) => {
+  const { trigger, isMutating } = useSWRMutation(
+    deactivate
+      ? "/api/manage/admin/deactivate-user"
+      : "/api/manage/admin/activate-user",
+    userMutationFetcher,
+    {
+      onSuccess: () => {
+        mutate();
+        setPopup({
+          message: `User ${deactivate ? "deactivated" : "activated"}!`,
+          type: "success",
+        });
+      },
+      onError: (errorMsg) =>
+        setPopup({ message: errorMsg.message, type: "error" }),
+    }
+  );
+  return (
+    <Button
+      className="w-min"
+      onClick={() => trigger({ user_email: user.email })}
+      disabled={isMutating}
+      size="sm"
+    >
+      {deactivate ? "Deactivate" : "Activate"}
+    </Button>
+  );
+};
